requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
beautifulsoup4==4.12.2
requests==2.31.0
feedparser==6.0.10
openai==1.12.0
python-dotenv==1.0.0
lxml==5.1.0
